+++ editor-github/package.json
@@ -10,6 +10,7 @@
   "keywords": [],
   "author": "",
   "license": "ISC",
+  "type": "module",
   "devDependencies": {
     "@google-cloud/functions-framework": "^3.1.2",
     "@octokit/auth-app": "^4.0.7",
